Added functions to save and restore global interface state, intended for use when a program needs to switch between mutually exclusive GUIs.
Added themes, allowing the definition of a set of defaults for window
appearance and behaviour.

Revamped the event system. Information such as x and y location and event type
is now encapsulated within <GUI-Event> class and its subclasses, which include
<Key-Event>, <Mouse-Event> etc.

Windows now have a slot, window-event-handler, which can be bound to a function
of the form (lambda (win event) boolean). If bound, it is called first if the
window is sent an event.

New functions: map-set-foreground-at, hide-all-windows

Added code to main gui loop to allow lisp process to respond to slime/swank
commands while the engine is running.