We now depend on the Alexandria library, available at: http://common-lisp.net/project/alexandria/
"start-gui" allows specification of renderer and fullscreen mode.
Default renderer is GLSL.

We no longer attempt to auto detect the desktop resolution, as the libtcod
function that does this no longer works (it gives the resolution of the root
console rather than the whole screen). If width and height are not supplied
when the root console is initialised, they default to 80 and 25.

Added README.md in Markdown format.